Fix Sonar issues: reduce break/continue statements, cast operands to long/double, use specialized functional interfaces

+++ 2024/src/main/java/info/jab/aoc2024/day15/WarehouseScaled.java
@@ -51,7 +51,7 @@ public class WarehouseScaled {
             .toList();
 
         String commands = inputLines.stream()
-            .skip(grid.size() + 1)
+            .skip((long) grid.size() + 1)
             .collect(Collectors.joining());
 
         return new InputData(grid, commands.toCharArray());
